🔨 refactor: display mode feature 분리
