+++ src/app/layout.tsx
@@ -4,7 +4,7 @@ import Script from "next/script";
 import { Toaster } from "@/src/components/ui/sonner";
 import RootProvider from "../contexts/RootProvider";
 import FCMNotification from "../components/fcm/FCMNotification";
-import { Providers } from "../components/display-mode/ThemeProvider";
+import { Providers } from "@/src/features/display-mode/display-mode-common/components/theme-provider";
 import ThemeColorUpdater from "../components/home/ThemeColorUpdater";
 import { getMetadata } from "../utils";
 export const metadata: Metadata = getMetadata();
