+++ src/features/display-mode/display-mode-common/components/theme-provider.tsx
